Move to new method for ajax updates
When a regular html request received, the page is rendered with
communities.haml. The response should be a complete page, and it should
be usable for every task the ajax version is (this is not complete yet).
When a request for 'application/json' is received, the page is rendered
with layout communities.json, and the html is put into the json layout
using content_for.

communities.json builds a json document that is a mapping of jQuery
selectors to html content. The merge function on the client handles
the actually replacement. If a value is false or null, nothing is done.